Refactor trump card to set no trump last round
This is definitely a kludge, the getTrumpCardAsString() method on the
Round class is sub-optimal. Ultimately, we should refactor this to make
each card use a Card object. Then a getTrumpCard() method could always
return a Card object (but in the case of no trump, that Card object
would not have a value - it would be a null Card).

For now, we'll leave with this solution.

Fix #1